+++ src/BeliefParticles.cpp
@@ -9,7 +9,7 @@ std::unordered_map<int64_t, std::vector<int64_t>> BeliefUpdate(
     const BeliefParticles& b, int64_t action, int64_t num_sims,
     SimInterface* pomdp) {
   std::unordered_map<int64_t, std::vector<int64_t>> next_beliefs;
-  for (int i = 0; i < num_sims; ++i) {
+  for (int64_t i = 0; i < num_sims; ++i) {
     const int64_t state = b.SampleOneState();
     const auto [sNext, obs, reward, done] = pomdp->Step(state, action);
     next_beliefs[obs].push_back(state);
